fix: correção de bug da entidade temporadas

+++ src/main/java/com/doback/E_rank/facade/TemporadasFacade.java
@@ -27,7 +27,7 @@ public class TemporadasFacade {
         temporadaApplication.criarTemporada(temporada);
     }
 
-    public void excluirTemporada(long id) {
+    public void excluirTemporada(int id) {
         temporadaApplication.excluirTemporada(id);
     }
